Everything works
So far eveything works as intended. You can play a full game. Although checking the README there are still a few tweaks to be made. The biggest addition would be to add an endgame screen report because I don't think the original game ever ended. Would need to go back to check.